+++ src/nodes/DruvaMSP/ReportUsage.node.execute.ts
@@ -116,8 +116,8 @@ export async function executeReportUsageOperation(
       // Set endpoint based on operation
       const endpoint =
         operation === 'getItemizedConsumption'
-          ? '/msp/v2/reports/consumption/itemized/v2'
-          : '/msp/v2/reports/quota/itemized/v2';
+          ? '/msp/reporting/v2/reports/consumptionItemized'
+          : '/msp/reporting/v2/reports/quotaItemized';
 
       // Create filters array for all filters
       const filterBy: IReportFilter[] = [];